option to list/kill

+++ stopRPi.pl
@@ -2,11 +2,17 @@
 use strict;
 use warnings;
 
+
 my @pidList;
 my @ProcList = ('SystemMonitor.pl','MapNetwork.pl','WatchDog.pl','ConfigServer.pl','ShutdownNUC.pl','i2c_apc_v6.pl',
 				'RemoteLog.pl', 'RunNVR.pl', 'RemoteData.pl','WatchSSD.pl','i2c_apc_v6.pl','Watcher.pl','WatchRPi.pl');
-my $ret = `ps -elf |grep perl`;
-print $ret;
+my $ret;
+if ($ARGV[0] eq 'kill') {
+	my $ret = `ps -elf |grep perl` ;
+	print $ret;
+} else {
+	print "Listing only, cmd line option 'kill' to kill all\n";
+}
 foreach my $curProc (@ProcList) {
 	my $return = `ps -C $curProc -o pid`;
  	my @lines = split("\n",$return);
@@ -18,11 +24,13 @@ foreach my $curProc (@ProcList) {
  	}
 	push(@pidList,$pid) unless $pid == -1;
 }
-if (scalar @pidList > 0) {
+if ( (scalar @pidList > 0) && ($ARGV[0] eq 'kill')) {
 	my $cmd = "kill " . join(" ",@pidList);
 	`$cmd`;	
+	$ret = `ps -elf |grep perl`;
+} else {
+	
 }
-$ret = `ps -elf |grep perl`;
 print $ret;
 
 1;
